Improve bazel error message
And move logic to a function so that variables can be closer to where they are used
